Fixed Bug - in register_serialize_entry.
Avevo scelto un tipo di dato unsigned per testare segnali di errore potenzialmente negativi.

+++ register.c
@@ -130,7 +130,7 @@ char* register_serialize_entry(const struct entry* E, char* buf, size_t len)
 {
     char str[ENTRY_TEXT_MAXLEN] = {};
     char* ans;
-    size_t strLen;
+    int strLen;
     char symbol;
 
     if (E == NULL)
